implementing scaling of theme
