+++ website/src/styles/Theme/colors.ts
@@ -1,0 +1,8 @@
+export const colors = {
+  light: {
+    bg: "gray.100"
+  },
+  dark: {
+    bg: "gray.900"
+  },
+};
